operations splitted
