+++ bpy_speckle/connector/operations/__init__.py
@@ -1,0 +1,2 @@
+from ..blender_operators.load_button import SPECKLE_OT_load  # noqa: F401
+from ..operations.load_operation import load_operation  # noqa: F401
